Add files via upload
In this lab we look at ways to visualize our data in order to construct multivariate regression models. We explore what information is provided to us when we estimate linear models in R, and we have a brief look at some model diagnostics. We also explore the usefulness of our model using test and training data methods.